Change GetMemoryOccupation return type to size_t

+++ include/CKBehaviorIO.h
@@ -93,7 +93,7 @@ public:
 
     virtual void PreDelete();
 
-    virtual int GetMemoryOccupation();
+    virtual size_t GetMemoryOccupation();
 
     //--------------------------------------------
     // Dependencies functions	{secret}
