+++ include/CKParameter.h
@@ -80,7 +80,7 @@ public:
 
     virtual void CheckPostDeletion();
 
-    virtual int GetMemoryOccupation();
+    virtual size_t GetMemoryOccupation();
     virtual int IsObjectUsed(CKObject *o, CK_CLASSID cid);
 
     //--------------------------------------------
