+++ src/CK2_3D/RCKAnimation.h
@@ -21,7 +21,7 @@ public:
     CKStateChunk *Save(CKFile *file, CKDWORD flags) override;
     CKERROR Load(CKStateChunk *chunk, CKFile *file) override;
 
-    int GetMemoryOccupation() override;
+    size_t GetMemoryOccupation() override;
 
     void CheckPreDeletion() override;
